Unit test improvement
This was the kind of prioritization that we agreed would make the most sense. In case of commits with equal weights, the newest commit prioritzed higher (because they're often the commit that introduced the change). Also, decreases are expressed with the same magnitude as increases (as seen with test commit 6).