Commit rerun prioritization (#15)
* Commit prioritization module

* Minor refactor

* Small optimization and improved algorithm

Favors spikes less.

* Unit test

* Unit test improvement

This was the kind of prioritization that we agreed would make the most sense. In case of commits with equal weights, the newest commit prioritzed higher (because they're often the commit that introduced the change). Also, decreases are expressed with the same magnitude as increases (as seen with test commit 6).

* Sort by most recent date

Now, in case of commits with equal priority (it happens sometimes but it's not super frequent), the newest commit is prioritized higher.

* New change formula

Expresses decreases with the same magnitude as increases. Whereas decreases were expressed by 0 to 1 and increases by 0 to inf, both are now unlimited.

* Fix test header

* Move script to modules/